fix: i64 全面修复——typeck_coerce_let_int_lit iv int→int64_t + AST_EXPR_LIT i64 推断
5 处 i64 截断修复（完整链路）：
1. lexer.c: lex_number int→int64_t + lex_hex_number 不截断
2. typeck.c: fold_expr int→int64_t（二元+一元 CTFE 运算）
3. typeck.c: AST_EXPR_LIT 超出 i32 范围自动推断为 i64
4. typeck.c: typeck_coerce_let_int_lit iv int→int64_t（u32 大常量不截断）
5. codegen.c: const_folded_val + AST_EXPR_LIT i64 用 %lld

验证：
- i64_arith: 1000000000000*2=2000000000000 ✅ EXIT=42
- i64_ctfe: INT64_MIN ✅ EXIT=42
- vmm.sx: u32 大常量 0xE4524FFE ✅ exit=0
- macOS 30 QEMU gate: 30/30 ✅

+++ compiler/src/typeck/typeck.c
@@ -936,7 +936,7 @@ static int typeck_let_init_matches(const struct ASTType *decl, const struct ASTE
  * let 初值：整数字面量按声明类型收窄/零初始化（与 typeck.sx 一致）。
  */
 static int typeck_coerce_let_int_lit(const struct ASTType *decl, struct ASTExpr *init) {
-    int iv;
+    int64_t iv;
     enum ASTTypeKind dk;
     if (!decl || !init || init->kind != AST_EXPR_LIT)
         return 0;
@@ -2162,7 +2162,6 @@ static int typeck_expr_sym(const struct ASTExpr *e, const char **names,
                 return typeck_expr_sym(e->value.unary.operand, names, type_kinds, type_names, n, type_ptrs, inside_loop, struct_defs, num_structs, enum_defs, num_enums, m);
             return 0;  /* return; 的 void 检查在 typeck_block 中根据 func_return_type 完成 */
         case AST_EXPR_LIT:
-            /* i32 范围外的整数字面量自动推断为 i64（避免大整数截断） */
             if (e->value.int_val > 2147483647LL || e->value.int_val < -2147483648LL)
                 ((struct ASTExpr *)e)->resolved_type = &static_type_i64;
             else
